Add files via upload
The schlieren uses concave mirrors on either side of the test section of the wind tunnel. A mercury vapor lamp or a spark gap system is used as a bright source of light. The light is passed through a slit which is placed such that the reflected light from the mirror forms parallel rays that pass through the test section. On the other side of the tunnel, the parallel rays are collected by another mirror and focused to a point at the knife edge. The rays continue on to a recording device like a video camera.
Python code is for Image pre-processing taken by video camera. 
CAD models for mirror stand and stand for camera alternative for tripod.
CAD model for light source and knife edge stand.